Refactor JS base classes
