+++ js/src/_base/Three.js
@@ -1,10 +1,7 @@
 var _ = require('underscore');
 var widgets = require("@jupyter-widgets/base");
-var pkgName = require('../../package.json').name;
 var Promise = require('bluebird');
 var $ = require('jquery');
-
-var RendererPool = require('./RendererPool');
 
 var Enums = require('./enums');
 
@@ -13,412 +10,6 @@ var ThreeCache = {
     byUuid: {},
     byName: {}
 };
-
-function extendedLog( object3D ) {
-    object3D.traverse(function(obj) {
-        var s = '|___';
-        var obj2 = obj;
-
-        while (obj2.parent !== null) {
-            s = '\t' + s;
-            obj2 = obj2.parent;
-        }
-        console.log(s + obj.name + ' <' + obj.type + '>');
-    });
-};
-
-var RenderableModel = widgets.DOMWidgetModel.extend({
-
-    defaults: function() {
-        return _.extend(widgets.DOMWidgetModel.prototype.defaults.call(this), {
-            _model_module: pkgName,
-            _view_module: pkgName,
-            _model_name: 'RenderableModel',
-            _view_name: 'RenderableView',
-
-            _width: 200,
-            _height: 200,
-        });
-    },
-
-});
-
-var RenderableView = widgets.DOMWidgetView.extend({
-
-    initialize: function () {
-        widgets.DOMWidgetView.prototype.initialize.apply(this, arguments);
-
-        // starts as "frozen" until renderer is acquired
-        this.isFrozen = true;
-        this.id = Math.floor(Math.random() * 1000000);
-    },
-
-    remove: function() {
-        widgets.DOMWidgetView.prototype.remove.apply(this, arguments);
-
-        if (!this.isFrozen) {
-            this.isFrozen = true;
-            RendererPool.release(this.renderer);
-            this.renderer = null;
-        }
-    },
-
-    render: function() {
-        this.doRender();
-    },
-
-    doRender: function() {
-        this.el.className = "jupyter-widget jupyter-threejs";
-
-        this.acquireRenderer();
-
-        this.lazyRendererSetup();
-
-        this.on('destroy', this.destroy, this);
-        this.listenTo(this.model, 'rerender',       this.renderScene);
-        this.listenTo(this.model, 'msg:custom',     this.onCustomMessage.bind(this));
-
-        this.listenTo(this.model, 'change:_width',  this.updateSize.bind(this));
-        this.listenTo(this.model, 'change:_height', this.updateSize.bind(this));
-    },
-
-    tick: function() {
-        requestAnimationFrame(this.tock.bind(this));
-    },
-
-    tock: function() {
-        this.renderScene();
-        if (this.animate) {
-            this.tick();
-        }
-    },
-
-    destroy: function() {
-        this.$el.empty();
-        if (!this.isFrozen) {
-            this.teardownViewer();
-        }
-    },
-
-    updateSize: function() {
-        var width = this.model.get('_width');
-        var height = this.model.get('_height');
-        this.renderer.setSize(width, height);
-    },
-
-    renderScene: function() {
-        this.log('renderScene');
-
-        // TODO: check renderer.domElement.isContextLost()
-
-        if (this.isFrozen) {
-            this.log('renderScene->isFrozen');
-
-            this.acquireRenderer();
-            this.updateSize();
-        }
-
-        this.renderer.render(this.scene, this.camera);
-    },
-
-    teardownViewer: function() {
-
-        this.$renderer.off('mouseenter');
-        this.$renderer.off('mouseleave');
-
-        this.isFrozen = true;
-        RendererPool.release(this.renderer);
-
-        this.$renderer = null;
-        this.renderer = null;
-
-        this.disableControls();
-
-        this.$el.css('margin-bottom', 'auto');
-
-    },
-
-    acquireRenderer: function() {
-        if (!this.isFrozen) {
-            return;
-        }
-        this.isFrozen = false;
-
-        this.log('ThreeView.acquiring...');
-
-        if(this.$frozenRenderer) {
-            this.$frozenRenderer.off('mouseenter');
-            this.$frozenRenderer = null;
-        }
-
-        this.renderer = RendererPool.acquire(this.onRendererReclaimed.bind(this));
-        this.renderer.setSize(this.model.get('_width'), this.model.get('_height'));
-
-        this.$renderer = $(this.renderer.domElement);
-        this.$el.empty().append(this.$renderer);
-
-        this.$el.css('margin-bottom', '-5px');
-
-        this.log('ThreeView.acquireRenderer(' + this.renderer.poolId + ')');
-
-        if (this.controls) {
-            this.enableControls();
-        }
-    },
-
-    freeze: function() {
-        if (this.isFrozen) {
-            this.log('already frozen...');
-            return;
-        }
-
-        this.log('ThreeView.freeze(id=' + this.renderer.poolId + ')');
-
-        this.animate = false;
-        this.$el.empty().append('<img src="' + this.renderer.domElement.toDataURL() + '" />');
-
-        this.teardownViewer();
-
-        this.$frozenRenderer = this.$el.find('img');
-        this.$frozenRenderer.on('mouseenter', _.bind(function() {
-            this.log('frozenRenderer.mouseenter');
-            this.tick(); // renderer will be acquired by renderScene
-        }, this));
-
-    },
-
-    enableControls: function() {
-        this.log('Enable controls');
-        var that = this;
-        this.controls.forEach(function(control) {
-            control.enabled = true;
-            control.connectEvents(that.$el[0]);
-            control.addEventListener('change', that.tick.bind(that));
-        });
-    },
-
-    disableControls: function() {
-        this.log('Disable controls');
-        var that = this;
-        this.controls.forEach(function(control) {
-            control.enabled = false;
-            control.dispose();  // Disconnect from DOM events
-            control.removeEventListener('change', that.tick.bind(that));
-        });
-    },
-
-    onCustomMessage: function(content, buffers) {
-        switch(content.type) {
-            case 'freeze':
-                this.freeze();
-                break;
-            default:
-        }
-    },
-
-    onRendererReclaimed: function() {
-        this.log('ThreeView WebGL context is being reclaimed: ' + this.renderer.poolId);
-        this.freeze();
-    },
-
-    log: function(str) {
-        console.log('TV(' + this.id + '): ' + str);
-    },
-
-    lazyRendererSetup: function() {
-        throw new Error('RenderableView should not be used directly, please subclass!');
-    }
-});
-
-
-var PreviewView = RenderableView.extend({
-
-    render: function() {
-        // ensure that model is fully initialized before attempting render
-        return this.model.initPromise.bind(this).then(this.doRender);
-    },
-
-    constructScene: function() {
-
-        var obj = this.model.get('child').obj;
-
-        this.clearScene();
-        this.scene.add(this.camera);
-
-        if (obj instanceof THREE.Object3D) {
-
-            this.log('render Object3D');
-            this.scene.add(obj);
-
-        } else if (obj instanceof THREE.Geometry || obj instanceof THREE.BufferGeometry) {
-
-            var material;
-            if (this.model.get('_flat')) {
-                material = new THREE.MeshPhongMaterial({
-                    color: '#ffffff',
-                    shading: THREE.FlatShading,
-                });
-            } else if (this.model.get('_wire') || obj instanceof THREE.WireframeGeometry) {
-                material = new THREE.MeshBasicMaterial({
-                    color: '#888888',
-                    wireframe: true,
-                    shading: THREE.FlatShading,
-                });
-            } else {
-                material = new THREE.MeshStandardMaterial({
-                    color: '#ffffff',
-                });
-            }
-
-            var mesh = new THREE.Mesh(obj, material);
-            this.scene.add(mesh);
-
-        } else if (obj instanceof THREE.Material) {
-
-            var geometry = new THREE.SphereGeometry(15, 16, 12);
-            var mesh = new THREE.Mesh(geometry, obj);
-            this.scene.add(mesh);
-
-        } else if (obj instanceof THREE.Texture) {
-
-            var geometry = new THREE.SphereGeometry(15, 16, 12);
-            var mat = new THREE.MeshStandardMaterial({ map: obj });
-            var mesh = new THREE.Mesh(geometry, mat);
-            this.scene.add(mesh);
-
-        }
-
-    },
-
-    clearScene: function() {
-        // this.controls.reset();
-        this.scene.children.forEach(function(child) {
-            this.scene.remove(child);
-        }, this);
-    },
-
-    update: function() {
-
-        RenderableView.prototype.update.apply(this, arguments);
-
-        if (this.model.get('child').obj) {
-            this.constructScene();
-        }
-        this.renderScene();
-
-    },
-
-    renderScene: function() {
-        this.log('renderScene');
-
-        // TODO: check renderer.domElement.isContextLost()
-
-        if (this.isFrozen) {
-            this.log('renderScene->isFrozen');
-
-            this.acquireRenderer();
-            this.updateSize();
-            this.enableControls();
-
-            if (this.model.get('child').obj) {
-                this.constructScene();
-            }
-        }
-
-        this.renderer.render(this.scene, this.camera);
-    },
-
-    setupControls: function() {
-        // Allow user to inspect object with mouse/scrollwheel
-        this.log('setting up controls');
-        var control = new THREE.OrbitControls(this.camera, this.renderer.domElement);
-        control.target.set(0, 0, 0);
-        control.update();
-        this.controls = [control];
-    },
-
-    lazyRendererSetup: function() {
-        this.camera = new THREE.PerspectiveCamera(60, 1.0); // aspect is updated by this.updateSize()
-        this.camera.position.set(-40, 40, 40);
-        this.camera.lookAt(new THREE.Vector3(0,0,0));
-
-        this.updateSize();
-
-        this.scene = new THREE.Scene();
-        // cameras need to be added to scene
-        this.scene.add(this.camera);
-
-        // Lights
-        this.pointLight = new THREE.PointLight('#ffffff', 1, 0);
-        this.pointLight.position.set(-100, 100, 100);
-        this.pointLight.lookAt(new THREE.Vector3(0,0,0));
-        this.ambLight = new THREE.AmbientLight('#ffffff', 0.5);
-        this.camera.add(this.ambLight);
-        this.camera.add(this.pointLight);
-
-        this.setupControls();
-        this.enableControls();
-
-        if (this.model.get('child').obj) {
-            this.constructScene();
-            this.renderScene();
-        }
-    },
-
-    updateSize: function() {
-        RenderableView.prototype.updateSize.call(this);
-        var width = this.model.get('_width');
-        var height = this.model.get('_height');
-        this.camera.aspect = width / height;
-        this.camera.updateProjectionMatrix();
-        var cameraModel = this.model.get('camera');
-        cameraModel.set({
-            aspect: cameraModel.obj.aspect,
-        });
-        this.tick();
-    },
-
-});
-
-
-var PreviewModel = RenderableModel.extend({
-
-    defaults: function() {
-        return _.extend(RenderableModel.prototype.defaults.call(this), {
-            _model_name: 'PreviewModel',
-            _view_name: 'PreviewView',
-
-            _flat: false,
-            _wire: false,
-            child: null,
-        });
-    },
-
-    initialize: function(attributes, options) {
-        RenderableModel.prototype.initialize.apply(this, arguments);
-
-        this.initPromise = this.get('child').initPromise.bind(this).then(function() {
-            this.setupListeners();
-        });
-    },
-
-    setupListeners: function() {
-
-        this.get('child').on('change', this.onChildChange, this);
-
-    },
-
-    onChange: function(model, options) {
-        this.trigger('rerender', this, {});
-    },
-
-}, {
-
-    serializers: _.extend({
-        child: { deserialize: widgets.unpack_models },
-    }, RenderableModel.serializers),
-
-});
 
 
 var ThreeModel = widgets.WidgetModel.extend({
@@ -1115,9 +706,5 @@ var ThreeModel = widgets.WidgetModel.extend({
 });
 
 module.exports = {
-    PreviewModel: PreviewModel,
-    PreviewView: PreviewView,
-    RenderableModel: RenderableModel,
-    RenderableView: RenderableView,
     ThreeModel: ThreeModel,
 };
